Revised Nanite System mechanics, tidied up code, added more HUD Sprites.
HUD Sprites have yet to be implemented. The Nanite System has been revised, mainly for the exit condition of Crisis Mode, and for making the Health Regen rate dependent on Nanite Pool and the presence of Berserk.

Tidied up code for the Armor definitions, and added replacement classes for the Berserk, Soulsphere and Megasphere.

Also added HUD Sprites in anticipation of implementing the Ammo/Weapons display system.